move profiling to general.py
