+++ schematic/utils/general.py
@@ -2,6 +2,9 @@
 from typing import Any, Dict, Optional, Text
 import os
 import math
+import pstats
+from cProfile import Profile
+from functools import wraps
 
 def find_duplicates(_list):
     """Find duplicate items in a list"""
@@ -70,3 +73,61 @@ def convert_gb_to_bytes(gb: int):
     """
     bytes_to_return = gb * 1024 * 1024 * 1024
     return bytes_to_return
+
+def profile(output_file=None, sort_by='cumulative', lines_to_print=None, strip_dirs=False):
+    """
+    The function was initially taken from: https://towardsdatascience.com/how-to-profile-your-code-in-python-e70c834fad89
+    A time profiler decorator.
+    Inspired by and modified the profile decorator of Giampaolo Rodola:
+    http://code.activestate.com/recipes/577817-profile-decorator/
+    Args:
+        output_file: str or None. Default is None
+            Path of the output file. If only name of the file is given, it's
+            saved in the current directory.
+            If it's None, the name of the decorated function is used.
+        sort_by: str or SortKey enum or tuple/list of str/SortKey enum
+            Sorting criteria for the Stats object.
+            For a list of valid string and SortKey refer to:
+            https://docs.python.org/3/library/profile.html#pstats.Stats.sort_stats
+        lines_to_print: int or None
+            Number of lines to print. Default (None) is for all the lines.
+            This is useful in reducing the size of the printout, especially
+            that sorting by 'cumulative', the time consuming operations
+            are printed toward the top of the file.
+        strip_dirs: bool
+            Whether to remove the leading path info from file names.
+            This is also useful in reducing the size of the printout
+    Returns:
+        Profile of the decorated function
+    """
+
+    def inner(func):
+        @wraps(func)
+        def wrapper(*args, **kwargs):
+            _output_file = output_file or func.__name__ + '.prof'
+            pr = Profile()
+            pr.enable()
+            retval = func(*args, **kwargs)
+            pr.disable()
+            pr.dump_stats(_output_file)
+
+            #if we are running the functions on AWS: 
+            if "SECRETS_MANAGER_SECRETS" in os.environ:
+                ps = pstats.Stats(pr)
+                # limit this to 30 line for now otherwise it will be too long for AWS log
+                ps.sort_stats('cumulative').print_stats(30)
+            else: 
+                with open(_output_file, 'w') as f:
+                    ps = pstats.Stats(pr, stream=f)
+                    if strip_dirs:
+                        ps.strip_dirs()
+                    if isinstance(sort_by, (tuple, list)):
+                        ps.sort_stats(*sort_by)
+                    else:
+                        ps.sort_stats(sort_by)
+                    ps.print_stats(lines_to_print)
+            return retval
+
+        return wrapper
+
+    return inner
